patch on the scanner for limiting the number of scan steps
The max value for the number of steps in a scan is defined in the config file.
If the user configure the scanner and reach more steps than this limit, the actuators positions calculation is not done and an error message will pop up when *set_scan* button is triggered.
this is made in order to limit memory use and crashs when incorrect step size is entered