patch on TCPServer Actuator
poll_moving was done on each side: server and client hence twice to much polling.
So I:
deactivated the polling server side
set to longer the default interval time for the polling (in config file) 50ms -> 200ms.

the TCPServer actuator plugin has no multiaxes options (yet) so had to patch the check on this in set_file_preset in the dashboard