roi filtering silent crash
In case a user select a ROI and move it around like creasy, the event loop went crowded to process and display lineouts.
So changed the emission of the ROI changed signal to ROI change finished one.
This means the graph and process update append when the user release (with its mouse) the ROI.